Moved tagged structures to separate testing file.
